Specialize 'struct sum' for T + S and bool
It was previously too generic (actually an implementation of
sum-of-products), which made it hard to switch over to type_info-based
GC. Therefore, I split it into its only two uses: type T + S and type
bool.

As a side effect, the debug printing knows what the constructor names
are again, which is nice.