WIP: Lists (Parser, CPS, RTS)
I realized that 'uncons' was hard to implement as a primitive, because
'list a -> () + (a, list a)' requires both nullary products and
polymorphic binary products.

So, I went back and baked 'uncons' into the syntax for case analysis on
lists. Clumsy, but a decent stopgap until I have general case
expressions.

Also, much easier to lower, because I can reuse case expression codegen.