Implement the "argument pad" calling convention
Reserve a chunk of bytes for the arguments. Use auxiliary 'struct' to
interpret that buffer as a bundle of arguments.

Simplify pretty much everything, from tracing the thunk to the RTS
interface to (sort of) the codegen.

Works cleanly for polymorphic code and instantiations. Rarely allocates.

Also lays a bit more groundwork for "Closed Thunk Types".
I am satisfied.