Use withParameterList, attempt withEnvDef
withEnvDef runs into problems because paying more attention to scoping
means that environment fields were getting renamed and desynchronized
between the EnvDecl and the EnvAlloc.

I believe the solution to this is to recognize that environment fields
are not Place:s in the same way as local value bindings. Instead, they
are much closer to being record field labels, which are not scoped.