WIP: Revamp emitValue
With the new record creation function, emitting a value allocation is an
expression, not a sequence of statements. So, defineLocal no longer
takes a list of field initializers.

I intended to do a follow-up simplification to make emitValue return a
C expression, but this was stymied by the fact that I need to know the
type of a ctor app to appropriately apply the argument casts.

Storing type arguments in a CtorApp would resolve this (and is more
rigorous), but I need to pipe information around (all the way back to
CPS, I think.)