Experiment: inlining
"Fast and Effective Procedure Inlining", by Kent Dybvig.

The algorithm is now mostly implemented; it needs to be tested.

The exception is letrec expressions, which are omitted from the
algorithm and only described in prose.

I also hope to refactor the algorithm, for simplicity and correctness,
and also to be more suitable for my language instead of the lisp-like
language used by the paper.