Add an optimization pass for inlining single-use continuations
That is, continuations which are statically used in a single location.
This is contrasted with a single-invocation continuation, which is
dynamically used only once. Consider this:

case x of { true -> k; false -> k }

Regardless of the value of 'x', 'k' will only be invoked once, so 'k'
could theoretically use a different calling convention or
implementation. However, there are two occurrences of the variable 'k',
meaning that inlining k's definition (as done in this optimization pass)
would lead to code duplication.