Backend: Add FieldRef Name:s
As I try to make the environment parameter more like a normal parameter,
having a dedicate construct for referencing it (EnvName) is maybe not
the best. Also, there are other cases where I want to reference some
field of an object, that aren't specifically "env" references.

Thus, a new constructor for Name that lets you project fields.